// src/types/IContainerDeploymentCreate.tsx
// SPDX-FileCopyrightText: 2024 Osnabrück University of Applied Sciences
// SPDX-FileContributor: Andreas Schliebitz
// SPDX-FileContributor: Henri Graf
// SPDX-FileContributor: Jonas Tüpker
// SPDX-FileContributor: Lukas Hesse
// SPDX-FileContributor: Maik Fruhner
// SPDX-FileContributor: Prof. Dr.-Ing. Heiko Tapken
// SPDX-FileContributor: Tobias Wamhof
//
// SPDX-License-Identifier: MIT

import IPortBinding from './IPortBinding';

export default interface IContainerDeploymentCreate {
    name: string;
    edge_device_id: number | '';
    container_image_id: number;
    runtime: number;
    port_bindings: Array<IPortBinding>;
}